Split classes in different files
